Got shiny text, updated summit page for 2025
