+++ src/components/rainbow_text.module.css
@@ -1,0 +1,17 @@
+@keyframes shine {
+    0% {
+        background-position: 0% 50%;
+    }
+
+    100% {
+        background-position: 200% 50%;
+    }
+}
+
+.shine {
+    background: linear-gradient(90deg, rgb(128, 0, 128), orange);
+    background-size: 200%;
+    -webkit-background-clip: text;
+    color: transparent;
+    animation: shine 5s linear infinite;
+}
