+++ tailwind.config.ts
@@ -21,12 +21,19 @@ const config: Config = {
         "sm-wai-gray": "0 1px 1px rgba(60, 60, 60, 0.25)",
         "lg-wai-gray": "0 20px 13px rgba(60, 60, 60, 0.25)",
       },
+      backgroundSize: {
+        '200%': '200%',
+      },
       keyframes: {
         colourPulse: {
           "0%, 100%": {
             "stop-color": "#3c3c3c",
           },
           "50%": { "stop-color": "#6D60C1" },
+        },
+        shine: {
+          '0%': { 'background-position': '0% 50%' },
+          '100%': { 'background-position': '200% 50%' },
         },
         text: {
           "0%, 100%": {
@@ -42,6 +49,7 @@ const config: Config = {
       animation: {
         colourPulse: "colourPulse 10s ease-in-out infinite",
         colourPulseOffset: "colourPulse 10s ease-in-out infinite -5s",
+        shine: 'shine 5s linear infinite',
         text: "text 1.5s ease-in-out infinite",
       },
     },
